+++ scripts/cmake.sh
@@ -2,4 +2,4 @@
 
 cmake -S /Users/allenzhou/Research_software/Blast_wave -B /Users/allenzhou/Research_software/Blast_wave/build
 
-cmake --build /Users/allenzhou/Research_software/Blast_wave/build
+cmake --build /Users/allenzhou/Research_software/Blast_wave/build -j 8